Backend/requirements.txt
fastapi
uvicorn
python-multipart
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
pillow